WIP: implement safer Rust PKCS#11 abstraction
Still a work-in-progress. Published for preliminary feedback.

This commit introduces a new Rust-level abstraction that is intended to
be safer than what is currently available. This should allow a user to
use the crate with memory safety guarantees, and without having to use
raw pointers.

Things that might still need doing:

* Remove cruft. If we decide to move ahead with this abstraction, there
  is probably still a few things (imports, tests, structs) that are no
  longer used and should perhaps be removed).

* Verify memory safety.

* Improve ergonomics.